Tasks Completed

+++ BackendDeveloperTest1/Test1/Controllers/AccountsController.cs
@@ -22,10 +22,12 @@ namespace Test1.Controllers
             _sessionFactory = sessionFactory;
         }
 
+        // GET: api/accounts
         [HttpGet]
         public async Task<ActionResult<IEnumerable<AccountDto>>> List(CancellationToken cancellationToken)
         {
-            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken).ConfigureAwait(false);
+            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken)
+                .ConfigureAwait(false);
             const string sql = @"
 SELECT
     Guid,
@@ -38,6 +40,7 @@ SELECT
     NextBillingUtc
 FROM account";
 
+            //SqlBuilder Unnecessary here
             var builder = new SqlBuilder();
 
             var template = builder.AddTemplate(sql);
@@ -49,10 +52,12 @@ FROM account";
             return Ok(rows);
         }
 
+        // GET: api/accounts/{Guid}
         [HttpGet("{id:Guid}")]
         public async Task<ActionResult<AccountDto>> GetById(Guid id, CancellationToken cancellationToken)
         {
-            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken).ConfigureAwait(false);
+            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken)
+                .ConfigureAwait(false);
 
             const string sql = @"
 SELECT
@@ -76,13 +81,15 @@ FROM account
                 Guid = id
             });
 
-            var account = await dbContext.Session.QueryFirstOrDefaultAsync<AccountDto>(template.RawSql, template.Parameters, dbContext.Transaction).ConfigureAwait(false);
+            var account = await dbContext.Session.QueryFirstOrDefaultAsync<AccountDto>(template.RawSql, template.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
 
             dbContext.Commit();
 
             return account == null ? NotFound() : Ok(account);
         }
 
+        // GET: api/accounts/{Guid}/members
         [HttpGet("{id:Guid}/members")]
         public async Task<ActionResult<IEnumerable<MemberDto>>> GetMembers(Guid id, CancellationToken cancellationToken)
         {
@@ -99,27 +106,32 @@ SELECT
     m.Cancelled
 FROM account a
 LEFT JOIN member m ON m.AccountUid = a.UID
-WHERE a.Guid = @Guid
+/**where**/;
 ";
 
             var builder = new SqlBuilder();
 
-            var template = builder.AddTemplate(sql, new
+            var template = builder.AddTemplate(sql);
+
+            builder.Where("a.Guid = @Guid", new
             {
                 Guid = id
             });
 
-            var rows = await dbContext.Session.QueryAsync<MemberDto>(template.RawSql, template.Parameters, dbContext.Transaction).ConfigureAwait(false);
+            var rows = await dbContext.Session.QueryAsync<MemberDto>(template.RawSql, template.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
 
             dbContext.Commit();
 
             return Ok(rows);
 
         }
-
+        
+        // POST: api/accounts
         [HttpPost]
         public async Task<ActionResult> Create([FromBody] CreateAccountDto model, CancellationToken cancellationToken)
         {
+            //transaction makes sense here
             await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken).ConfigureAwait(false);
 
             //Validate
@@ -130,6 +142,7 @@ WHERE a.Guid = @Guid
                 return BadRequest(error);
             }
 
+            //ensure location exists and fetch location Uid
             const string getLocationSql = @"
 SELECT UID
 FROM location
@@ -140,10 +153,11 @@ WHERE Guid = @Guid;
             var getLocationTemplate = locationBuilder.AddTemplate(getLocationSql, new { Guid = model.LocationGuid });
             var location = await dbContext.Session.QueryFirstOrDefaultAsync<int?>(getLocationTemplate.RawSql, getLocationTemplate.Parameters, dbContext.Transaction).ConfigureAwait(false);
 
+            //Rollback if location was not found
             if (location == null)
             {
                 dbContext.Rollback();
-                return NotFound("Location not found.");
+                return NotFound("Location not found");
             }
 
             const string sql = @"
@@ -178,7 +192,7 @@ INSERT INTO account (
                 LocationUid = location.Value,
                 Guid = newGuid,
                 CreatedUtc = DateTime.UtcNow,
-                Status = AccountStatusType.GREEN,
+                Status = model.Status,
                 AccountType = model.AccountType,
                 PaymentAmount = model.PaymentAmount,
                 PendCancel = 0,
@@ -189,11 +203,19 @@ INSERT INTO account (
 
             var count = await dbContext.Session.ExecuteAsync(sql, parameters, dbContext.Transaction).ConfigureAwait(false);
 
-            dbContext.Commit();
+            //Rollback if create failed
+            if(count != 1)
+            {
+                dbContext.Rollback();
+                return BadRequest("Unable to add account");
+            }
 
-            return count == 1 ? StatusCode(StatusCodes.Status201Created, new { id = newGuid }) : BadRequest("Unable to add account");
+            dbContext.Commit();
+            //return new account Guid to test with postman
+            return StatusCode(StatusCodes.Status201Created, new { id = newGuid });
         }
 
+        // UPDATE: api/accounts/{Guid}
         [HttpPut("{id:Guid}")]
         public async Task<ActionResult> UpdateById(Guid id, [FromBody] UpdateAccountDto model, CancellationToken cancellationToken)
         {
@@ -207,6 +229,7 @@ INSERT INTO account (
                 return BadRequest(error);
             }
 
+            //model represents state change, use patch for updating specific fields
             const string sql = @"
 UPDATE account
 SET
@@ -217,7 +240,7 @@ SET
     PendCancel = @PendCancel,
     PendCancelDateUtc = @PendCancelDateUtc,
     EndDateUtc = @EndDateUtc
-WHERE Guid = @Guid";
+WHERE Guid = @Guid"; //Builder could be used for additional filters
 
             var builder = new SqlBuilder();
 
@@ -233,43 +256,74 @@ WHERE Guid = @Guid";
                 EndDateUtc = model.EndDateUtc,
             });
 
-            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction).ConfigureAwait(false);
+            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
 
             dbContext.Commit();
-
-            return count == 1 ? NoContent() : NotFound();
+            
+            if(count == 1)
+                return NoContent();
+            else
+                return NotFound();
         }
 
+        // DELETE: api/accounts/{Guid}
         [HttpDelete("{id:Guid}")]
         public async Task<ActionResult> DeleteById(Guid id, CancellationToken cancellationToken)
         {
-            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken).ConfigureAwait(false);
+            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken)
+                .ConfigureAwait(false);
 
+            //ensure acount exists, get keys
+            const string accountKeySql = @"SELECT UID FROM account WHERE Guid = @Guid";
+
+            var accountKeyBuilder = new SqlBuilder();
+            var accountKeyTemplate = accountKeyBuilder.AddTemplate(accountKeySql, new { Guid = id });
+            var account = await dbContext.Session.QueryFirstOrDefaultAsync<int?>(accountKeyTemplate.RawSql, accountKeyTemplate.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
+
+            //rollback if account not found
+            if(account == null)
+            {
+                dbContext.Rollback();
+                return NotFound("Account not found");
+            }
+
+            //delete members due to forieign key constraint
+            const string deleteMemberSql = @"DELETE FROM ""member"" WHERE AccountUid = @AccountUid";
+
+            var deleteMemberBuilder = new SqlBuilder();
+            var deleteMemberTemplate = deleteMemberBuilder.AddTemplate(deleteMemberSql, new {AccountUid = account.Value});
+            await dbContext.Session.ExecuteAsync(deleteMemberTemplate.RawSql, deleteMemberTemplate.Parameters, dbContext.Transaction);
+
+            //delete account
             const string sql = "DELETE FROM account WHERE Guid = @Guid;";
 
             var builder = new SqlBuilder();
+            var template = builder.AddTemplate(sql, new { Guid = id });
+            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
 
-            var template = builder.AddTemplate(sql, new
+
+            //rollback if account not delete
+            if (count != 1)
             {
-                Guid = id
-            });
-
-            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction).ConfigureAwait(false);
+                dbContext.Rollback();
+                return BadRequest("Unable to delete account");
+            }
 
             dbContext.Commit();
-
-            if (count == 1)
-                return Ok();
-            else
-                return BadRequest("Unable to delete location");
-
+            return Ok();
         }
 
+        // DELETE: api/accounts/{Guid}/members
         [HttpDelete("{id:Guid}/members")]
         public async Task<ActionResult> DeleteMembersByAccountId(Guid id, CancellationToken cancellationToken)
         {
-            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken).ConfigureAwait(false);
+            await using var dbContext = await _sessionFactory.CreateContextAsync(cancellationToken)
+                .ConfigureAwait(false);
 
+            //Delete non primary members that belong to this account 
             const string sql = @"
 DELETE FROM member 
 WHERE EXISTS (
@@ -278,7 +332,7 @@ WHERE EXISTS (
     WHERE a.Guid = @Guid 
         AND a.UID = member.AccountUid 
         AND member.""Primary"" = 0
-);";
+);";// we can use builder for additional filters in the future
 
             var builder = new SqlBuilder();
             var template = builder.AddTemplate(sql, new
@@ -286,10 +340,11 @@ WHERE EXISTS (
                 Guid = id
             });
 
-            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction).ConfigureAwait(false);
+            var count = await dbContext.Session.ExecuteAsync(template.RawSql, template.Parameters, dbContext.Transaction)
+                .ConfigureAwait(false);
 
             dbContext.Commit();
-            return NoContent();
+            return Ok();
         }
     }
 }
